Refactor adaptive memory test loader
